PI-815: Moved src/Hooks/Deactivation to src/WordPress/Hooks.
