+++ src/WordPress/WordPressAPI.php
@@ -73,7 +73,7 @@ class WordPressAPI implements IntegrationAPIInterface
     /**
      * We wrap the return value with an array to be consistent between
      * other plugins.
-     * 
+     *
      * @param null $userId
      *
      * @return mixed
@@ -142,20 +142,5 @@ class WordPressAPI implements IntegrationAPIInterface
         }
 
         return false;
-    }
-
-    public function clearDataStore()
-    {
-        $pluginKeys = \CF\API\Plugin::getPluginSettingsKeys();
-
-        // Delete Plugin Setting Options
-        foreach ($pluginKeys as $optionName) {
-            $this->dataStore->clear($optionName);
-        }
-
-        // Delete DataStore Options
-        $this->dataStore->clear(DataStore::API_KEY);
-        $this->dataStore->clear(DataStore::EMAIL);
-        $this->dataStore->clear(DataStore::CACHED_DOMAIN_NAME);
     }
 }
